Add paintball color key
