+++ src/Colormap.ts
@@ -167,195 +167,5 @@ function makeTextureImage(colormap: ColorMap) {
     return cmap_image;
 }
 
-type ColorbarOrientation = 'horizontal' | 'vertical';
-type ColorbarTickDirection = 'top' | 'bottom' | 'left' | 'right';
-interface ColorBarOptions {
-    /** The label to place along the color bar */
-    label?: string;
-
-    /** 
-     * An array of numbers to use as the tick locations. 
-     * @default Use all the levels in the color map provided to {@link makeColorBar}.
-     */
-    ticks?: number[];
-
-    /** 
-     * The direction the ticks should face. Valid values are 'left' and 'right' if orientation is 'vertical' and 'top' and 
-     * 'bottom' if orientation is 'horizontal'.
-     * @default 'left' if orientation is 'vertical' and 'bottom' if orientation is 'horizontal'
-     */
-    tick_direction?: ColorbarTickDirection;
-
-    /** 
-     * The orientation for the color bar. Valid values are 'horizontal' and 'vertical'.
-     * @default 'vertical'
-     */
-    orientation?: ColorbarOrientation;
-
-    /** 
-     * A font face to use for the label and tick values.
-     * @default 'sans-serif'
-     */
-    fontface?: string;
-};
-
-/**
- * Make an SVG containing a color bar. The color bar can either be oriented horizontal or vertical, and a label can be provided.
- * @param colormap - The color map to use
- * @param opts     - The options for creating the color bar
- * @returns An SVGElement containing the color bar image.
- * @example
- * // Create the color bar
- * const svg = makeColorBar(color_map, {label: 'Wind Speed (kts)', orientation: 'horizontal', 
- *                                      fontface: 'Trebuchet MS'});
- * 
- * // Add colorbar to the page
- * document.getElementById('colorbar-container').appendChild(svg);
- */
-function makeColorBar(colormap: ColorMap, opts: ColorBarOptions) {
-    const label = opts.label || "";
-    const ticks = opts.ticks || colormap.levels;
-    const orientation = opts.orientation || 'vertical';
-    const fontface = opts.fontface || 'sans-serif';
-
-    const tick_dir = opts.tick_direction || (orientation == 'vertical' ? 'left' : 'bottom');
-
-    if (orientation == 'vertical' && (tick_dir == 'top' || tick_dir == 'bottom') ||
-        orientation == 'horizontal' && (tick_dir == 'left' || tick_dir == 'right')) {
-        throw `tick_direction of '${tick_dir} doesn't match an orientation of ${orientation}`;
-    }
-
-    const createElement = (tagname: string, attributes?: Record<string, string | number>, parent?: SVGElement) => {
-        const elem = document.createElementNS('http://www.w3.org/2000/svg', tagname);
-
-        if (attributes !== undefined) {
-            Object.entries(attributes).forEach(([k, v]) => {
-                elem.setAttribute(k, v.toString());
-            });
-        }
-
-        if (parent !== undefined) {
-            parent.appendChild(elem);
-        }
-
-        return elem;
-    };
-
-    const getNChar = (n: number) => {
-        return Math.floor(Math.log10(Math.abs(n))) + (n < 0 ? 2 : 1);
-    }
-
-    const chars_left = getNChar(ticks[0]);
-    const chars_right = getNChar(ticks[ticks.length - 1]);
-
-    const bar_long_size = 600;
-    const bar_cross_size = bar_long_size / 9;
-    const bar_long_pad = orientation == 'horizontal' ? Math.max(chars_left, chars_right) * 4.5 : 5;
-    const bar_cross_pad = 3;
-    const bar_thickness = 10;
-
-    let height: number, width: number, bar_left: number, bar_top: number, bar_width: number, bar_height: number;
-
-    if (orientation == 'vertical') {
-        height = bar_long_size;
-        width = bar_cross_size;
-
-        bar_left = tick_dir == 'left' ? bar_cross_size - bar_cross_pad - bar_thickness : bar_cross_pad;
-        bar_top = bar_long_pad;
-        bar_width = bar_thickness;
-        bar_height = bar_long_size - 2 * bar_long_pad;
-    }
-    else {
-        width = bar_long_size;
-        height = bar_cross_size - 6;
-
-        bar_left = bar_long_pad;
-        bar_top = tick_dir == 'bottom' ? bar_cross_pad : bar_cross_size - 6 - bar_cross_pad - bar_thickness;
-        bar_height = bar_thickness;
-        bar_width = bar_long_size - 2 * bar_long_pad;
-    }
-
-    const n_colors = colormap.colors.length;
-
-    const root = createElement('svg', {width: width, height: height});
-    const gbar = createElement('g', {}, root);
-
-    let gtickattrs;
-    if (orientation == 'vertical') {
-        gtickattrs = tick_dir == 'left' ? {'text-anchor': 'end', transform: `translate(${bar_left}, ${bar_top})`} : 
-                                          {transform: `translate(${bar_left + bar_width}, ${bar_top})`}
-    }
-    else {
-        gtickattrs = tick_dir == 'bottom' ? {'text-anchor': 'middle', transform: `translate(${bar_left}, ${bar_top + bar_height})`} : 
-                                            {'text-anchor': 'middle', transform: `translate(${bar_left}, ${bar_top})`}
-    }
-    const gticks = createElement('g', gtickattrs, root);
-
-    colormap.colors.forEach((color, icolor) => {
-        let attrs: Record<string, string | number> = {};
-
-        if (orientation == 'vertical') {
-            attrs = {
-                x: bar_left, y: bar_top + bar_height * (1 - (icolor + 1) / n_colors), width: bar_width, height: bar_height / n_colors};
-        }
-        else {
-            attrs = {x: bar_left + bar_width * icolor / n_colors, y: bar_top, width: bar_width / n_colors, height: bar_height};
-        }
-
-        createElement('rect', {...attrs, fill: color.color, opacity: color.opacity}, gbar);
-    });
-
-    ticks.forEach(level => {
-        const ilevel = colormap.levels.indexOf(level);
-        const tickattrs = orientation == 'vertical' ? {transform: `translate(0, ${bar_height * (1 - ilevel / n_colors)})`} : 
-                                                      {transform: `translate(${bar_width * ilevel / n_colors}, 0)`};
-        const gtick = createElement('g', tickattrs, gticks);
-
-        let lineattrs;
-        if (orientation == 'vertical') {
-            lineattrs = tick_dir == 'left' ? {x2: -6} : {x2: 6};
-        }
-        else {
-            lineattrs = tick_dir == 'bottom' ? {y2 : 6} : {y2: -6};
-        }
-
-        createElement('line', {...lineattrs, stroke: '#000000', 'stroke-width': 1.5}, gtick);
-
-        let textattrs;
-        if (orientation == 'vertical') {
-            textattrs = tick_dir == 'left' ? {x: -9, dy: '0.32em'} : {x: 9, dy: '0.32em'};
-        }
-        else {
-            textattrs = tick_dir == 'bottom' ? {y: 9, dy: '0.8em'} : {y: -9, dy: '0em'};
-        }
-
-        const text = createElement('text', {...textattrs, fill: '#000000', style: `font-family: ${fontface};`}, gtick);
-        text.textContent = level.toString();
-    });
-
-    const outline_attrs = {
-        x: bar_left,
-        y: bar_top,
-        width: bar_width,
-        height: bar_height,
-        stroke: '#000000',
-        'stroke-width': 1.5,
-        fill: 'none'
-    };
-    createElement('rect', outline_attrs, root);
-
-    let labelattrs;
-    if (orientation == 'vertical') {
-        labelattrs = tick_dir == 'left' ? {transform: `translate(15, ${height / 2}) rotate(-90)`} : {transform: `translate(${width - 6}, ${height / 2}) rotate(-90)`};
-    }
-    else {
-        labelattrs = tick_dir == 'bottom' ? {transform: `translate(${width / 2}, ${height - 5})`} : {transform: `translate(${width / 2}, 15)`}
-    }
-    const label_elem = createElement('text', {...labelattrs, fill: '#000000', 'text-anchor': 'middle', style: `font-family: ${fontface};`}, root);
-    label_elem.textContent = label;
-
-    return root;
-}
-
-export {ColorMap, bluered, redblue, pw_speed500mb, pw_speed850mb, pw_cape, pw_t2m, pw_td2m, makeColorBar, makeTextureImage}
-export type {Color, ColorbarOrientation, ColorbarTickDirection, ColorBarOptions};
+export {ColorMap, bluered, redblue, pw_speed500mb, pw_speed850mb, pw_cape, pw_t2m, pw_td2m, makeTextureImage}
+export type {Color};
